perf(core): pack the ids and links an instruction carries
An `Instruction` was 160 bytes, a `Mnemonic` 80 and an operand 16, and a
session's body on /usr/bin/bash holds two million instructions. Now 112,
56 and 8:

- The module-interned ids (`LiteralId`, `VarnodeId`, `BytesId`,
  `PoisonId`) and the body-local temporary ids are `u32`, as the other
  body-local ids already were; a `LocalValueId` is eight bytes.
- The links an instruction carries — its block, its neighbours, its
  first use — and its machine address are packed (`value::link`): `None`
  is the one pattern no arena issues. Reads and writes go through
  `Option`, serde sees the same `Option`, and `WithUsers` sets a head
  instead of lending it.
- The argument lists of terminators, calls, intrinsics, p-code ops and
  tuples are `Box<[LocalValueId]>` rather than `Vec`: a mnemonic never
  grows one in place, and the box is a third smaller.

Peak memory of a session over bash: 695 → 560 MB. No change to the IR.

+++ core/src/value/bytes.rs
@@ -25,7 +25,7 @@ use crate::{
 use jstd::Identifier;
 
 #[derive(Identifier)]
-pub struct BytesId(usize);
+pub struct BytesId(u32);
 
 /// A compile-time opaque byte blob stored in a [`Context`](crate::context::Context).
 ///
